services/ml/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
scikit-learn>=1.5.0
numpy>=1.26.0
kafka-python>=2.0.2
httpx>=0.27.0
python-dotenv>=1.0.0